# Conflicts:
#	src/crate-writer.cc
#	src/crate-writer.hh
#	src/next/parser/ascii-parser-prim.cc
#	src/stage-converter.cc
#	tests/run-usdcat-compare.sh
#	tools/tusdchecker/CMakeLists.txt
#	web/npm/package-lock.json
#	web/npm/scripts/stage-package.mjs

+++ examples/tusdview/parametric_tess.cc
@@ -240,7 +240,7 @@ int AdaptiveTessellator::computeTessLevel(const ParametricPrim& prim,
   else segments = 32;
 
   // Apply user quality scale
-  segments = std::max(3, static_cast<int>(segments * qualityScale));
+  segments = std::max(3, static_cast<int>(static_cast<float>(segments) * qualityScale));
   segments = std::min(segments, 64);  // cap at 64
 
   // Ensure minimum for different primitive types
